WIP: Some refactoring of submodules
