+++ src/Operations.jl
@@ -8,7 +8,7 @@ import LibGit2
 
 import REPL
 using REPL.TerminalMenus
-using ..Types, ..GraphType, ..Resolve, ..PlatformEngines, ..GitTools, ..Display
+using ..Types, ..Resolve, ..PlatformEngines, ..GitTools, ..Display
 import ..depots, ..depots1, ..devdir, ..Types.uuid_julia, ..Types.PackageEntry
 import ..Artifacts: ensure_all_artifacts_installed, artifact_names
 using ..BinaryPlatforms
@@ -294,7 +294,7 @@ function resolve_versions!(ctx::Context, pkgs::Vector{PackageSpec})
     fixed[uuid_julia] = Fixed(VERSION)
     graph = deps_graph(ctx, names, reqs, fixed)
     simplify_graph!(graph)
-    vers = resolve(graph)
+    vers = Resolve.resolve(graph)
 
     find_registered!(ctx, collect(keys(vers)))
     # update vector of package versions
@@ -407,7 +407,7 @@ function deps_graph(ctx::Context, uuid_to_name::Dict{UUID,String}, reqs::Require
         end
     end
 
-    return Graph(all_versions, all_deps, all_compat, uuid_to_name, reqs, fixed, #=verbose=# ctx.graph_verbose)
+    return Resolve.Graph(all_versions, all_deps, all_compat, uuid_to_name, reqs, fixed, #=verbose=# ctx.graph_verbose)
 end
 
 function load_urls(ctx::Context, pkgs::Vector{PackageSpec})
